requirements.txt
appdirs==1.4.4
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
DateTime==4.3
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pipenv==2020.11.15
psycopg2-binary==2.8.6
pybase64==1.1.1
pytesseract==0.3.6
pytz==2020.4
regex==2020.11.13
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.20
tesseract==0.1.3
textblob==0.15.3
tqdm==4.54.1
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zope.interface==5.2.0
